pyproject.toml
[project]
name = "adversarial-attacks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "matplotlib>=3.10.0",
    "numpy>=2.2.0",
    "pillow>=11.0.0",
    "torch>=2.5.1",
    "torchvision>=0.20.1",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
